move kgraph to ngraph, separate file.
